+++ sdk/src/integrationTest/java/AccountBalanceIntegrationTest.java
@@ -17,6 +17,7 @@ import java.util.Objects;
 import java.util.concurrent.TimeUnit;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
@@ -24,16 +25,12 @@ class AccountBalanceIntegrationTest {
     @Test
     @DisplayName("can connect to previewnwet with TLS")
     void canConnectToPreviewnetWithTLS() throws Exception {
-        var network = new HashMap<String, AccountId>();
-        network.put("0.previewnet.hedera.com:50212", new AccountId(3));
-        network.put("1.previewnet.hedera.com:50212", new AccountId(4));
-        network.put("2.previewnet.hedera.com:50212", new AccountId(5));
-        network.put("3.previewnet.hedera.com:50212", new AccountId(6));
-        network.put("4.previewnet.hedera.com:50212", new AccountId(7));
+        var client = Client.forPreviewnet()
+            .setTransportSecurity(true);
 
-        var client = Client.forNetwork(network);
+        for (var entry : client.getNetwork().entrySet()) {
+            assertTrue(entry.getKey().endsWith(":50212"));
 
-        for (var entry : network.entrySet()) {
             new AccountBalanceQuery()
                 .setNodeAccountIds(Collections.singletonList(entry.getValue()))
                 .setAccountId(entry.getValue())
@@ -45,20 +42,40 @@ class AccountBalanceIntegrationTest {
     @Test
     @DisplayName("can connect to testnet with TLS")
     void canConnectToTestnetWithTLS() throws Exception {
-        var network = new HashMap<String, AccountId>();
-        network.put("0.testnet.hedera.com:50212", new AccountId(3));
-        network.put("1.testnet.hedera.com:50212", new AccountId(4));
-        network.put("2.testnet.hedera.com:50212", new AccountId(5));
-        network.put("3.testnet.hedera.com:50212", new AccountId(6));
-        network.put("4.testnet.hedera.com:50212", new AccountId(7));
+        var client = Client.forTestnet()
+            .setTransportSecurity(true);
 
-        var client = Client.forNetwork(network);
+        for (var entry : client.getNetwork().entrySet()) {
+            assertTrue(entry.getKey().endsWith(":50212"));
 
-        for (var entry : network.entrySet()) {
             new AccountBalanceQuery()
                 .setNodeAccountIds(Collections.singletonList(entry.getValue()))
                 .setAccountId(entry.getValue())
                 .execute(client);
+        }
+
+        client.close();
+    }
+
+    @Test
+    @DisplayName("can connect to previewnet with certificate verification off")
+    void cannotConnectToPreviewnetWhenNetworkNameIsNullAndCertificateVerificationIsEnabled() throws Exception {
+        var client = Client.forPreviewnet()
+            .setTransportSecurity(true)
+            .setVerifyCertificates(true)
+            .setNetworkName(null);
+
+        assertFalse(client.getNetwork().isEmpty());
+
+        for (var entry : client.getNetwork().entrySet()) {
+            assertTrue(entry.getKey().endsWith(":50212"));
+
+            assertThrows(IllegalStateException.class, () -> {
+                new AccountBalanceQuery()
+                    .setNodeAccountIds(Collections.singletonList(entry.getValue()))
+                    .setAccountId(entry.getValue())
+                    .execute(client);
+            });
         }
 
         client.close();