+++ sdk/src/main/java/com/hedera/hashgraph/sdk/TopicMessageQuery.java
@@ -144,7 +144,13 @@ public final class TopicMessageQuery {
     public SubscriptionHandle subscribe(Client client, Consumer<TopicMessage> onNext) {
         SubscriptionHandle subscriptionHandle = new SubscriptionHandle();
         HashMap<TransactionID, ArrayList<ConsensusTopicResponse>> pendingMessages = new HashMap<>();
-        makeStreamingCall(client, subscriptionHandle, onNext, 0, new AtomicLong(), new AtomicReference<>(), pendingMessages);
+
+        try {
+            makeStreamingCall(client, subscriptionHandle, onNext, 0, new AtomicLong(), new AtomicReference<>(), pendingMessages);
+        } catch (InterruptedException e) {
+            throw new RuntimeException(e);
+        }
+
         return subscriptionHandle;
     }
 
@@ -156,7 +162,7 @@ public final class TopicMessageQuery {
         AtomicLong counter,
         AtomicReference<ConsensusTopicResponse> lastMessage,
         HashMap<TransactionID, ArrayList<ConsensusTopicResponse>> pendingMessages
-    ) {
+    ) throws InterruptedException {
         // TODO: check status of channel before using it?
         ClientCall<ConsensusTopicQuery, ConsensusTopicResponse> call =
             client.mirrorNetwork.getNextMirrorNode().getChannel()
@@ -247,7 +253,11 @@ public final class TopicMessageQuery {
                     Thread.currentThread().interrupt();
                 }
 
-                makeStreamingCall(client, subscriptionHandle, onNext, attempt + 1, counter, lastMessage, pendingMessages);
+                try {
+                    makeStreamingCall(client, subscriptionHandle, onNext, attempt + 1, counter, lastMessage, pendingMessages);
+                } catch (InterruptedException e) {
+                    throw new RuntimeException(e);
+                }
             }
 
             @Override