+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v2.3.0
+
+### Added
+
+ * Support for toggling TLS for both mirror network and services network
+
 ## v2.2.0
 
 ### Added
@@ -38,6 +44,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
  * Added keep alive timeout of 10 seconds to all gRPC connections
+
+### Added
+
+* `Client.setTransportSecurity()` - Enable/Disable TLS for any node
 
 ### Changed